feat: add solutions modules tasks
